requirements.txt
aiogram==2.24.0
google-generativeai==1.0.0  # Replace with the correct version if needed
gunicorn==20.1.0
